[#2] Fix issue where log level is not respected
Log level and log format were not respected because imports
(specifically torcpy) set the log config before GitSlice.  To fix this,
we could use the force option but will instead choose to use the
coloredlogs packages which adds colour and includes some additional log
variables such as the hostname which is very useful to see which of the
nodes that log line came from.

Closes #2.