[#4 & #5] Support different containers, images and config options
As GitSlice traverses through commits the configuration options which
work to run a newer version may not work correctly on an older version -
 for example, if the requirements.txt file for Python applications is
moved or because an older version of the container (i.e. Python 3.10 to
3.9) is required.  It is now possible to set configuration options based
on the history of a commit.

Closes #4

Additional options for filtering commits have been added.  This is done
through two new configuration options.

Firstly, arguments can be passed to git-rev-list, for example, this
allows for selection of:

- Commits only by a particular author(s)
- Commits before, or after a particular date
- Ignore merge commits (i.e. commits with more than one parent)

Some additional options have also been added that are not provided
though git-rev-list, these are provided by GitSlice.  These include:

- Commits in which more or less than a threshold of changes occurred
  (i.e. thresholds for total files changes, additions and deletions).
- Minimum time between commits (referred to a Min Delta).  For example,
  it is possible to select one commit every 24 hours.

Closes #5.